introduce separate permission for backlogs mgmt
