+++ modules/backlogs/config/locales/en.yml
@@ -156,6 +156,7 @@ en:
 
   permission_view_master_backlog: "View master backlog"
   permission_view_taskboards: "View taskboards"
+  permission_select_done_statuses: "Select done statuses"
   permission_update_sprints: "Update sprints"
 
   points_accepted: "points accepted"
