+++ modules/backlogs/lib/open_project/backlogs/engine.rb
@@ -51,12 +51,6 @@ module OpenProject::Backlogs
              author_url: 'https://www.openproject.org',
              bundled: true,
              settings: settings do
-      OpenProject::AccessControl.permission(:edit_project).tap do |add|
-        add.controller_actions << 'projects/project_done_statuses'
-        add.controller_actions << 'projects/rebuild_positions'
-        add.controller_actions << 'backlogs_project_settings/show'
-      end
-
       OpenProject::AccessControl.permission(:add_work_packages).tap do |add|
         add.controller_actions << 'rb_stories/create'
         add.controller_actions << 'rb_tasks/create'
@@ -73,7 +67,8 @@ module OpenProject::Backlogs
         # SYNTAX: permission :name_of_permission, { :controller_name => [:action1, :action2] }
 
         # Master backlog permissions
-        permission :view_master_backlog, rb_master_backlogs: :index,
+        permission :view_master_backlog,
+                   rb_master_backlogs: :index,
                    rb_sprints: %i[index show],
                    rb_wikis: :show,
                    rb_stories: %i[index show],
@@ -81,7 +76,8 @@ module OpenProject::Backlogs
                    rb_burndown_charts: :show,
                    rb_export_card_configurations: %i[index show]
 
-        permission :view_taskboards, rb_taskboards: :show,
+        permission :view_taskboards,
+                   rb_taskboards: :show,
                    rb_sprints: :show,
                    rb_stories: :show,
                    rb_tasks: %i[index show],
@@ -90,10 +86,20 @@ module OpenProject::Backlogs
                    rb_burndown_charts: :show,
                    rb_export_card_configurations: %i[index show]
 
+        permission :select_done_statuses,
+                   {
+                     'projects/settings/backlogs': %i[show update rebuild_positions]
+                   },
+                   require: :member
+
         # Sprint permissions
         # :show_sprints and :list_sprints are implicit in :view_master_backlog permission
-        permission :update_sprints, rb_sprints: %i[edit update],
-                   rb_wikis: %i[edit update]
+        permission :update_sprints,
+                   {
+                     rb_sprints: %i[edit update],
+                     rb_wikis: %i[edit update],
+                   },
+                   require: :member
       end
 
       menu :project_menu,
@@ -113,7 +119,6 @@ module OpenProject::Backlogs
                Status
                Type
                Project
-               ProjectsController
                User
                VersionsController
                Version]
